backend/requirements.txt
# TerraMind Pre-Sowing Advisor — Backend Requirements
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
slowapi>=0.1.9
pydantic>=2.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
openpyxl>=3.1
xlrd>=2.0
python-multipart>=0.0.6
requests>=2.31.0
httpx>=0.27.0
networkx>=3.2
python-dotenv>=1.0.1
